container/user-import/requirements.txt
requests
pandas
python-keycloak
Unidecode